refactor: move inspect envelope contracts to output crate
